fix: Unsafe CSS modification prevents script init
Other extensions may insert `CSSStyleSheet`s into the page. When KS tries to insert its own CSS rules, it did so in an unsafe way that would conflict with those modifications, resulting in an `Failed to execute 'insertRule' on 'CSSStyleSheet': Cannot access StyleSheet to insertRule` error.

We now use a distinct `style` element for our own styles to work around this conflict.